fix: Field invoice.
