+++ src/components/ADempiere/FieldDefinition/FieldSearch/BusinessPartnerInfo/index.vue
@@ -74,7 +74,7 @@ import { RECORD_ROWS_BY_LIST } from '@/utils/ADempiere/dictionary/field/lookups'
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
 
 export default {
-  name: 'BusinessPartnerInfo',
+  name: 'BusinessPartnerInfoField',
 
   components: {
     ButtonBusinessPartnersList
