+++ src/components/ADempiere/FieldDefinition/FieldSearch/ProductInfo/index.vue
@@ -77,7 +77,7 @@ import { RECORD_ROWS_BY_LIST } from '@/utils/ADempiere/dictionary/field/lookups'
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
 
 export default {
-  name: 'BusinessPartnerInfo',
+  name: 'ProductInfoField',
 
   components: {
     ButtonProductsList
